Conditionally include jasmine based on source file
This is kind of a hack to work around the fact that boot.js doesn't have
any tests that it should run. Also, if we include boot.js from src, we
end up running the test suite twice, which is almost the exact opposite
of what we want to do (which is not run any tests if there aren't any).

[Finish #54582638]