Jack up the CSS and slide a new foundation under it
* Removed Flexbox compatibility mixins for old browsers
* Removed compatibility shims for IE <= 8
* Removed upstream Docco CSS in favor of our own small stylesheet
* Made Docco CSS opt-in, not opt-out
* Removed most of the reset stylesheet

This eliminates a three-way specificity fight between the reset
stylesheet, Docco, and everything else, which has been the cause of many
bugs and maintainability problems. The Docco CSS was designed to "own"
the page that it's in and never played well with others. The reset
stylesheet may have been a good solution to 2015's browser compatibility
problems. But it forces extra complexity on the other CSS rules, and in
2023 it creates more bugs than it prevents.

This is an appearance-preserving refactor with one major exception: code
examples are now dark text on a light background. Supporting dark mode
throught the site is a goal, but until that happens, it's better to be
consistent. Mixing light and dark mode on the same page like we did
before is the worst of both worlds from an accessibility standpoint.