Replaced expando header with two different fixed size headers
The shrink-as-you-scroll effect is cool, but it made the CSS difficult
to maintain. In particular, adding new layouts was harder than it needed
to be. Using a big header for the homepage and the docs index and a
small sticky header for other pages achieves most of the same goals
without as much complexity.