Rephrased browser/Node support FAQs to reflect the likelihood of dropping outdated versions in minor releases
We've long reserved the right to drop outdated platforms in minor releases,
but the frequency of major releases of jasmine-core meant that we rarely
needed to. The old wording of the FAQs reflected that. With no 6.0 release
of jasmine-core planned for the forseeable future, it's more likely that
outdated platform versions will be dropped in a minor releases.